Continue splitting MajorRequirementList
